+++ test/boost/UUID_test.cc
@@ -11,7 +11,7 @@
 #include <boost/test/unit_test.hpp>
 #include <utility>
 #include "utils/UUID_gen.hh"
-#include "types/types.hh"
+#include "utils/lexicographical_compare.hh"
 
 BOOST_AUTO_TEST_CASE(test_generation_of_name_based_UUID) {
     auto uuid = utils::UUID_gen::get_name_UUID("systembatchlog");
